refactor: provide worker implementation instead of patch global
